requirements.txt
asgiref==3.7.2
Brotli==1.0.9
dj-database-url==2.1.0
Django==4.2.4
django-ckeditor==6.7.0
django-js-asset==2.1.0
gunicorn==21.2.0
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
